ex 4.21

+++ part4/blog_list/controllers/blogs.js
@@ -61,8 +61,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    return response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    //const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    if ( blog.user.toString() === user._id.toString() ) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    }
+
+    return response.status(401).json({ error: 'invalid user' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
